Only add "iconv" to the .pc file if needed
On some systems, e.g. recent versions of Debian GNU/Linux,
the iconv(3) function is part of the system libc and there is
no need to use libiconv as a separate library. The AM_ICONV
macro detects that and leaves the LIBICONV variable empty;
in that case, there is no need to specify "iconv" as a private
dependency in the pkg-config file (and it is actually harmful,
since at least in Debian GNU/Linux there is no package that
provides an iconv.pc file).