7zip reader: add ARM64 filter test for deflate and other non-lzma compressors
Since liblzma provides its own BCJ filters, they need to be reimplemented
for other compressors. This adds a test for deflate, and is expected to fail
until the following commit adds support. Once support is added, it should
work for other compressors too.

This test archive contains a linux C hello world executable, built for
arm64, and created like so:
7zz a -t7z -m0=deflate -mf=ARM64 libarchive/test/test_read_format_7zip_deflate_arm64.7z hw-arm64