xar: replace UTF8Toisolat1 with a local reimplementation
In xar, Non-Latin-1 gets stored Base64-encoded. However:

- `UTF8Toisolat1` is a `libxml2` export.
- The writer code made a copy of every filename for the purposes of
	checking whether it was Latin-1 stored as UTF-8.

Instead of moving it into an abstraction with the rest of the libxml2
imports, I have elected to replace it with a simple UTF-8 validator.

The new function has slightly different results than the one it is
replacing for invalid UTF-8. Namely, it treats illegal UTF-8 "overlong"
encodings of Latin-1 codepoints as _invalid_. It operates on the
principle that we can determine whether something is Latin-1 based
entirely on how long the sequence is expected to be.

In practice, this does not matter because the xar writer passes all
filenames through a UTF-8 converter that rejects overlong encodings
before they get here.

Therefore, this replacement produces the same results as UTF8Toisolat1
with one fewer import and one fewer copy per filename.