Allow setting charset for archive.
This exposes archive.current_code so that the caller can explicitly specify a charset instead of relying on the default locale.

This is important for platforms like Android and iOS (and when compiled without iconv), because nl_langinfo isn't available before Android API 26, and on iOS nl_langinfo always returns US-ASCII. And the caller should be able to specify the charset they want to work with for each archive, instead of relying on a global state.

This change is updating the ZIP writing support to respect this per-archive charset when setting the UTF-8 flag as well.

Bug: #1572