Add archive_read_open_filenames_w, based on archive_read_open_filenames
There is a discrepancy between the w and non-w filename APIs, wherein a
consumer of libarchive can open a multi-volume set with names in the
current locale (on Windows) but not with UTF-16 names.

This patch addresses that issue.

archive_read_open_filename_w now delegates its work to
archive_read_open_filenames_w.

Fixes #1728