Parse tar headers incrementally
This rebuilds the tar reader to parse all header
data incrementally as it appears in the stream.

This definitively fixes a longstanding issue with unsupported
pax attributes.  Libarchive must limit the amount of data that
it reads into memory, and this has caused problems with large
unknown attributes.  By scanning iteratively, we can instead
identify an attribute and decide whether to read it into memory or
whether to skip it without reading.

This design also allows us to vary our sanity limits for
different pax attributes (e.g., an attribute that is a single number
can be limited to a few dozen bytes while an attribute holding an ACL is
allowed to be a few megabytes).  This allows us to be a little more
resistent to malicious archives that might try to force allocation of
very large amounts of memory, though there is still work to be done
here.

This builds on earlier work to make tar header parsing be purely
iterative by keeping the _first_ appearance of any given value instead
of the original architecture that recursively cached data in memory
in order to effectively process all the data from back-to-front.