Fix error handling around strtol() usages
The code here had a couple of bad code patterns that
seem to have been copied throughout:
* Checking errno after strtol() -- Standard C doesn't seem
  to actually require this, so we shouldn't rely on it
* Casting the result of strtol() directly to `int`.
  This loses information prematurely.

Instead, I've added `l` as a temporary of type `long`,
use that to hold the result of `strtol()` until it
can be checked.  I've also removed the `errno` tests
in favor of checking the end pointer value.