Prevent usage of uninitialized variable
Calling __archive_mkstemp can lead to access of an uninitialized
variable in __archive_mktempx, because temp_name is only initialized
if supplied template argument is NULL.

If template is not NULL, it is eventually compared with temp_name.s
anyway.

The fix is simple: Always initialize temp_name, which merely sets
values in the struct. No memory allocation occurs and the check leads
to the expected result.