7zip: Limit amount of properties
The uint64_t variable propertiesSize is eventually casted to size_t
which, on 32 bit systems, can result in integer truncation.

In such a situation, it is possible that less than the minimum of 5
properties are parsed and processed, which will result in out of
boundary reads in init_decompression because the error check
`if (coder1->propertiesSize < 5)` still takes the uint64_t variable
into account.