Turn counter into flag
When working with matches, the code does not care about the actual
amount of times when it matched, but just if it matched so far at
least once or never.

Turning the counter into a boolean flag has the advantage that even
insanely huge archives will never lead to integer overflow here.