rpm: Clarify meaning of magic number 16
The minimum header size is used at multiple instances in the code.
Clarify its meaning by using a definition.

Also, while in an if-block which checked if rpm->pos is 16, use the
value to avoid a warning with Visual Studio and to clarify that
more data will follow beginning at the current position.

Fixes:

warning C4244: '=': conversion from 'uint64_t' to 'long', possible loss of data