Rework tar linkname parsing
Instead of keeping a local copy of the linkname, push
it into the entry directly when we see it.  This requires
modifying archive_entry to store a single "linkname" value
and separately mark whether that name is a symlink or hardlink.
This should not break any existing use, since no entry should
ever have both a symlink and a hardlink value.

For tar, this lets us correctly handle PAX link names, which
are stored early in the archive and only later do we find out
whether that name is a symlink or hardlink.  (We did this before
by storing the link name and putting it into the entry only
after we knew whether this was a hard or sym link.)

Nice side effect:  This slightly reduces the size of the
archive_entry struct.

Still TODO:  I need to correctly handle character set
for link names.