Support ISOs with a non-standard PVD layouts
The CSRG ISOs have a non-standard PVD layout with a 68-byte
root directory record (rather than the 34-byte record required
by ECMA119/ISO9660).  I built a test image with this change
and modified the ISO9660 reader to accept it.

While I was working on the bid logic to recognize PVDs,
I added a number of additional correctness checks that should
make our bidding a bit more accurate.  In particular, this
should more than compensate for the weakened check of the root
directory record size.

Resolves #2232