Prevent overruning the `bidder` array
The `bidder` is assigned to point to the first element of a->bidders
before the for-loop.
Thus we want to start the `i` increment from 1, so it doesn't overrun
the array on the last increment of the for-loop which will be accessed
if the for-loop ends due to the unmet condition in the next `bidder->name` right after the for-loop.