Prevent possible use-after-free
The `entry` and `spare_entry` can actually point to the same memory
which could lead to a use-after-free undefined behaviour in C.

This can occur in the `archive_entry_linkify` function if the switch
case `ARCHIVE_ENTRY_LINKIFY_LIKE_NEW_CPIO` is true. Both entries can
point to the same `le->entry` causing the described behaviour.