Fix a minor date-parsing bug and fill in missing ISO9660 testing
This is somewhat academic, since we don't actually expose any of the
ISO9660 header information that is stored in 17-byte date format,
but inspection revealed an off-by-one error in the parsing here.

This also proved a nice motivation to fill in some verification
in our most basic ISO9660 test case.