Reject tar entries > 2^60 bytes
In particular, this ensures that we cannot overflow rounding-up
calculations.  Recent tar changes put in a lot of sanity limits
on the sizes of particular kinds of data, but the usual behavior
in most cases was to skip over-large values.  The skipping behavior
required rounding-up and accumulating values that could potentially overflow
64-bit integers.  This adds some coarser checks that fail more
directly when an entry claims to be more than 1 exbibyte (2^60 bytes),
avoiding any possibility of numeric overflow along these paths.

OSS-Fuzz Issue: 70062