autoconf: look for zlib >= 1.2.1
zlib 1.2.0 added this improvement for inflate:
"Raw inflate no longer needs an extra dummy byte at end"

libarchive does not feed zlib extra data beyond end of stream, so
it does not work with zlib < 1.2.0. However, it is difficult to
test for zlib 1.2.0, and easy to test for 1.2.1. Both are very
old, so let's just require zlib >= 1.2.1.

https://github.com/libarchive/libarchive/issues/615