Clarify crc32 variable names
No functional change, just a tiny style improvement.

Use `crc32_computed` to refer to the crc32 that the reader has
computed and `crc32_read` to refer to the value that we read from
the archive.  That hopefully makes this code a tiny bit easier to
follow.  (It confused me recently when I was double-checking something
in this area, so I thought an improvement here might help others.)