Reintroduce size_t cast to clarify the use of a 32b shift on a 64b arg
Building for 64-bit Windows with MSVC generates warning `C4334`, `'<<':
result of 32-bit shift implicitly converted to 64 bits (was 64-bit shift
intended?)`

The clearest way to signal to the compiler that we intend this to fit in
a `size_t` is to cast it.

We could also use a 64-bit shift with a 64-bit operand (e.g. `1ULL`,
which will be 64-bit on all platforms), but that will generate narrowing
warnings on 32-bit `size_t` platforms.