Verify reading of a GNU tar posix archive with 8GB member
POSIX says that we have to ignore the ustar size field when
there's a size specified in the pax extension header.  This
bug wasn't caught earlier because libarchive always puts
a correct size in both places, so the failure to ignore the
ustar size made no difference.  GNU tar follows the POSIX
spec a little more strictly, which triggers this bug.

Verifies our fix for #2404.