tar: fix bug when -s/a/b/ used more than once with b flag
When the -s/regexp/replacement/ option was used with the b flag more than
once, the result of the previous substitution was appended to the previous
subject instead of replacing it. Fixed it by making sure the subject is
made the empty string before the call to realloc_strcat(). That in effect
makes it more like a realloc_strcpy(), but creating a new realloc_strcpy()
function for that one usage doesn't feel worth it.

Resolves Issue libarchive/libarchive#2414