Polish for GNU tar format reading/writing
A few small tweaks to improve reading/writing of the
legacy GNU tar format.

* Be more tolerant of redundant 'K' and 'L' headers
* Fill in missing error messages for redundant headers
* New test for reading archive with redundant 'L' headers
* Earlier identification of GNU tar format in some cases

These changes were inspired by Issue #2434.  Although that was
determined to not technically be a bug in libarchive, it's relatively
easy for libarchive to tolerate duplicate 'K' and 'L' headers and
we should be issuing appropriate error messages in any case.