Issue 2548: Reading GNU sparse entries
My attempt to fix #2404 just made the confusion between the size
of the extracted file and the size of the contents in the tar
archive worse than it was before.

@ferivoz in #2557 showed that the confusion stemmed from a point
where we were setting the size in the entry (which is by definition
the size of the file on disk) when we read the `GNU.sparse.size`
and `GNU.sparse.realsize` attributes (which might represent the
size on disk or in the archive) and then using that to determine
whether to read the value in ustar header (which represents
the size of the data in the archive).

The confusion stems from two issues:
* The GNU.sparse.* fields mean different things depending
  on the version of GNU tar used.
* The regular Pax `size` field overrides the value in the
  ustar header, but the GNU sparse size fields don't always
  do so.
* The previous libarchive code tried to reconcile different
  size information as we went along, which is problematic
  because the order in which this information appears can
  vary.

This PR makes one big structural change:  We now have separate
storage for every different size field we might encounter.
We now just store these values and record which one we saw.
Then at the end, when we have all the information available
at once, we can use this data to determine the size on disk
and the size in the archive.

A few key facts about GNU sparse formats:

* GNU legacy sparse format:  Stored all the relevant info
  in an extension of the ustar header.
* GNU pax 0.0 format: Used `GNU.sparse.size` to store
  the size on disk
* GNU pax 0.1 format: Used `GNU.sparse.size` to store
  the size on disk
* GNU pax 1.0 format: Used `GNU.sparse.realsize` to store
  the size on disk; repurposed `GNU.sparse.size` to store
  the size in the archive, but omitted this in favor of
  the ustar size field when that could be used.

And of course, some key precedence information:
* Pax `size` field always overrides the ustar header size field.
* GNU sparse size fields override it ONLY when they
  represent the size of the data in the archive.